+++ source/extensions/omni.isaac.lab/omni/isaac/lab/utils/math.py
@@ -746,7 +746,7 @@ def is_identity_pose(pos: torch.tensor, rot: torch.tensor) -> bool:
     return torch.allclose(pos, pos_identity) and torch.allclose(rot, rot_identity)
 
 
-# @torch.jit.script
+@torch.jit.script
 def combine_frame_transforms(
     t01: torch.Tensor, q01: torch.Tensor, t12: torch.Tensor | None = None, q12: torch.Tensor | None = None
 ) -> tuple[torch.Tensor, torch.Tensor]: